% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_schema.R
\name{validate_schema}
\alias{validate_schema}
\title{Validates a JSON schema against the JSON schema specification}
\usage{
validate_schema(
  schema,
  version = c("draft202012", "draft201909", "draft07", "draft06", "draft04"),
  ...
)
}
\arguments{
\item{schema}{A JSON schema}

\item{version}{The version of the JSON schema specification to use}

\item{...}{Additional arguments passed to \code{\link[=validate]{validate()}}}
}
\value{
TRUE if the schema is valid, FALSE otherwise
}
\description{
Sources for the schema definitions can be found here:
\itemize{
\item \url{https://json-schema.org/draft/2020-12/schema}
\item \url{https://json-schema.org/draft/2019-09/schema}
\item \url{https://json-schema.org/draft-07/hyper-schema}
\item \url{https://json-schema.org/draft-06/hyper-schema}
\item \url{https://json-schema.org/draft-04/hyper-schema}
}
}
\examples{
validate_schema(schema = '{"type": "string"}', version = "draft202012")
}
\seealso{
\code{\link[=validate]{validate()}}
}
